Module X86_target

Declarations of some x86 targets.

val parent : Bap_core_theory.Theory.Target.t

The parent of all x86 targets.

When a new target is declared it is advised to use this target as parent so that the newly declared target will be included into the x86 Targets family. The parent target is pure abstract and doesn't have any propreties set.

val i86 : Bap_core_theory.Theory.Target.t

val i186 : Bap_core_theory.Theory.Target.t

val i286 : Bap_core_theory.Theory.Target.t

val i386 : Bap_core_theory.Theory.Target.t

val i486 : Bap_core_theory.Theory.Target.t

val i586 : Bap_core_theory.Theory.Target.t

val i686 : Bap_core_theory.Theory.Target.t

val amd64 : Bap_core_theory.Theory.Target.t

val load : unit -> unit

load () loads the knowledge base rules for the x86 targets.

This includes parsing the loader output and enabling backward compatibility with the old Arch.t representation.